Title: 특정 문자로 시작하는 문자열, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240412/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.cpp
@@ -23,7 +23,7 @@ int main() {
             sum += arr[i].length();
         } 
     }
-    avg = sum / cnt;
+    avg = (double)sum / cnt;
 
     cout << fixed;
     cout.precision(2);
